chore: update platform config
As far as I understand buildx, we need to use the BUILDPLATFORM for building stuff and the TARGETPLATFORM for the final image.
I noticed this when trying to build the sshd image locally (macOS) for a linux/amd64 target, it threw an exec format error on the
amd64 machine, which should not have happened.

This did not happen in the CI since the CI uses QEMU.

I think this fixes this issue.