chore: remove mailtrain
This image was built for mailtrain 1.x, which is EOL for a while.
Therefore, we also remove this image from the source.

Existing published versions will be kept.

This also adds a check if the directory exists - if it does not, we skip the build checks.
This is necessary to trigger builds on any changes, even deletions, but not when the whole
directory is deleted.